Remove RdfResourceFactory interface, rename class
