+++ tests/Event/ReadModel/RDF/EventJsonToTurtleConverterTest.php
@@ -14,7 +14,7 @@ use CultuurNet\UDB3\Model\ValueObject\Moderation\WorkflowStatus;
 use CultuurNet\UDB3\RDF\JsonDataCouldNotBeConverted;
 use CultuurNet\UDB3\RDF\NodeUri\CRC32HashGenerator;
 use CultuurNet\UDB3\RDF\NodeUri\NodeUriGenerator;
-use CultuurNet\UDB3\RDF\NodeUri\ResourceFactory\RdfResourceFactoryWithoutBlankNodes;
+use CultuurNet\UDB3\RDF\NodeUri\ResourceFactory\RdfResourceFactory;
 use CultuurNet\UDB3\ReadModel\DocumentRepository;
 use CultuurNet\UDB3\ReadModel\InMemoryDocumentRepository;
 use CultuurNet\UDB3\ReadModel\JsonDocument;
@@ -96,7 +96,7 @@ class EventJsonToTurtleConverterTest extends TestCase
             $this->documentRepository,
             (new EventDenormalizer())->handlesDummyOrganizers(),
             $addressParser,
-            new RdfResourceFactoryWithoutBlankNodes(new NodeUriGenerator(new CRC32HashGenerator())),
+            new RdfResourceFactory(new NodeUriGenerator(new CRC32HashGenerator())),
             new VideoNormalizer([]),
             $this->imageNormalizer,
             $this->logger
