+++ tests/Place/ReadModel/RDF/PlaceJsonToTurtleConverterTest.php
@@ -19,7 +19,7 @@ use CultuurNet\UDB3\Model\ValueObject\Moderation\WorkflowStatus;
 use CultuurNet\UDB3\RDF\JsonDataCouldNotBeConverted;
 use CultuurNet\UDB3\RDF\NodeUri\CRC32HashGenerator;
 use CultuurNet\UDB3\RDF\NodeUri\NodeUriGenerator;
-use CultuurNet\UDB3\RDF\NodeUri\ResourceFactory\RdfResourceFactoryWithoutBlankNodes;
+use CultuurNet\UDB3\RDF\NodeUri\ResourceFactory\RdfResourceFactory;
 use CultuurNet\UDB3\ReadModel\DocumentRepository;
 use CultuurNet\UDB3\ReadModel\InMemoryDocumentRepository;
 use CultuurNet\UDB3\ReadModel\JsonDocument;
@@ -102,7 +102,7 @@ class PlaceJsonToTurtleConverterTest extends TestCase
             $this->documentRepository,
             new PlaceDenormalizer(),
             $addressParser,
-            new RdfResourceFactoryWithoutBlankNodes(new NodeUriGenerator(new CRC32HashGenerator())),
+            new RdfResourceFactory(new NodeUriGenerator(new CRC32HashGenerator())),
             $this->logger
         );
     }
